fix(timer): fix systick current value mask

+++ src/arch/armv8-m/timer.c
@@ -1,3 +1,3 @@
 #include <timer.h>
 
-unsigned long TIMER_FREQ;
+unsigned long long TIMER_FREQ;
